fix: extract separate component for response emojis to prevent rerendering whole response card
