+++ packages/client/components/TeamPrompt/TeamPromptResponseEmojis.tsx
@@ -1,0 +1,64 @@
+import graphql from 'babel-plugin-relay/macro'
+import styled from '@emotion/styled'
+import {useFragment} from 'react-relay'
+import {TeamPromptResponseEmojis_response$key} from '~/__generated__/TeamPromptResponseEmojis_response.graphql'
+import useAtmosphere from '../../hooks/useAtmosphere'
+import useMutationProps from '../../hooks/useMutationProps'
+import AddReactjiToReactableMutation from '../../mutations/AddReactjiToReactableMutation'
+import ReactjiId from '../../shared/gqlIds/ReactjiId'
+import React from 'react'
+import ReactjiSection from '../ReflectionCard/ReactjiSection'
+
+const StyledReactjis = styled(ReactjiSection)({
+  paddingRight: '8px',
+  paddingTop: '8px'
+})
+
+interface Props {
+  meetingId: string
+  responseRef: TeamPromptResponseEmojis_response$key
+}
+
+export const TeamPromptResponseEmojis = (props: Props) => {
+  const {responseRef, meetingId} = props
+
+  const response = useFragment(
+    graphql`
+      fragment TeamPromptResponseEmojis_response on TeamPromptResponse {
+        id
+        reactjis {
+          id
+          count
+          isViewerReactji
+          ...ReactjiSection_reactjis
+        }
+      }
+    `,
+    responseRef
+  )
+  const {reactjis} = response
+  const atmosphere = useAtmosphere()
+
+  const {onError, onCompleted, submitMutation, submitting} = useMutationProps()
+
+  const onToggleReactji = (emojiId: string) => {
+    if (submitting || !response) return
+    const isRemove = !!reactjis.find((reactji) => {
+      return reactji.isViewerReactji && ReactjiId.split(reactji.id).name === emojiId
+    })
+    submitMutation()
+    AddReactjiToReactableMutation(
+      atmosphere,
+      {
+        reactableId: response?.id,
+        reactableType: 'RESPONSE',
+        isRemove,
+        reactji: emojiId,
+        meetingId
+      },
+      {onCompleted, onError}
+    )
+  }
+
+  return <StyledReactjis reactjis={reactjis} onToggle={onToggleReactji} />
+}
